Removed -ansi and added -fsigned-char to CFLAGS Add need explict cast in parser.c to avoid warning Explicitly set tabbing and tabular in those environments latex russian.tex before testing
git-svn-id: https://latex2rtf.svn.sourceforge.net/svnroot/latex2rtf/trunk@265 e562ef82-0334-0410-8083-dbe22b1e9a75

+++ parser.c
@@ -840,10 +840,10 @@ getDimension(void)
 	
 /* obtain unit of measure */
 	skipSpaces();
-	buffer[0] = tolower(getTexChar());
+	buffer[0] = tolower((int) getTexChar());
 	
 	if (buffer[0] != '\\') {
-		buffer[1] = tolower(getTexChar());
+		buffer[1] = tolower((int) getTexChar());
 		buffer[2] = '\0';
 		
 		diagnostics(4,"getDimension() dimension is <%s>", buffer);
